+++ NEWS.md
@@ -1,3 +1,12 @@
+# dropR (development version)
+
+- Install via `devtools::install_github("iscience-kn/dropR")`
+
+# dropR 1.0.4
+
+* Added new `do_print()` function for pretty output printing
+* Calculation for `add_dropout_idx()` and `compute_stats()` have been refined to allow more accuracy
+
 # dropR 1.0.3
 
 * CRAN resubmission after big package re-vamp!